+++ tests/Rule/NestedTest.php
@@ -277,37 +277,6 @@ final class NestedTest extends TestCase
     public function dataHandler(): array
     {
         return [
-            'object' => [
-                new class () {
-                    #[Nested(['number' => new Number(max: 7)])]
-                    private ObjectWithDifferentPropertyVisibility $object;
-
-                    public function __construct()
-                    {
-                        $this->object = new ObjectWithDifferentPropertyVisibility();
-                    }
-                },
-                [
-                    'object.number' => ['Value must be no greater than 7.'],
-                ],
-            ],
-            'object-private-only' => [
-                new class () {
-                    #[Nested(
-                        ['age' => new Number(min: 100, skipOnEmpty: true), 'number' => new Number(max: 7)],
-                        propertyVisibility: ReflectionProperty::IS_PRIVATE
-                    )]
-                    private ObjectWithDifferentPropertyVisibility $object;
-
-                    public function __construct()
-                    {
-                        $this->object = new ObjectWithDifferentPropertyVisibility();
-                    }
-                },
-                [
-                    'object.number' => ['Value must be no greater than 7.'],
-                ],
-            ],
             'class-string-rules' => [
                 new class () {
                     #[Nested(ObjectWithDifferentPropertyVisibility::class)]
